+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `--out` and `--xy` for plotting corresponding data files used with `--plot`.
+- `--surfaces` for showing surfaces settings.
 
 ### Removed